[TestNG] Deprecate not mapping TestNG tests to scenarios.
Previously TestNG tests could only either be mapped to the whole
cucumber suite, or to each feature in the cucumber suite. With the
possibility of mapping TestNG tests to each scenario, these other
options can be deprecated.
