--- examples/java-calculator-testng/src/test/java/cucumber/examples/java/calculator/RunCukesByScenarioAndCompositionTest.java
@@ -1,40 +1,0 @@
-package cucumber.examples.java.calculator;
-
-import cucumber.api.CucumberOptions;
-import cucumber.api.testng.CucumberFeatureWrapper;
-import cucumber.api.testng.PickleEventWrapper;
-import cucumber.api.testng.TestNGCucumberRunner;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeClass;
-import org.testng.annotations.DataProvider;
-import org.testng.annotations.Test;
-
-/**
- * An example of using TestNG when the test class does not inherit from 
- * AbstractTestNGCucumberTests but still executes each scenario as a separate
- * TestNG test.
- */
-@CucumberOptions(format = "json:target/cucumber-report-feature-composite.json")
-public class RunCukesByScenarioAndCompositionTest extends RunCukesByCompositionBase {
-    private TestNGCucumberRunner testNGCucumberRunner;
-
-    @BeforeClass(alwaysRun = true)
-    public void setUpClass() throws Exception {
-        testNGCucumberRunner = new TestNGCucumberRunner(this.getClass());
-    }
-
-    @Test(groups = "cucumber", description = "Runs Cucumber Feature", dataProvider = "scenarios")
-    public void scenario(PickleEventWrapper pickleEvent, CucumberFeatureWrapper cucumberFeature) throws Throwable {
-        testNGCucumberRunner.runScenario(pickleEvent.getPickleEvent());
-    }
-
-    @DataProvider
-    public Object[][] scenarios() {
-        return testNGCucumberRunner.provideScenarios();
-    }
-
-    @AfterClass(alwaysRun = true)
-    public void tearDownClass() throws Exception {
-        testNGCucumberRunner.finish();
-    }
-}
